# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProcessMiner API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tenants": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Create Tenant",
        "operationId": "create_tenant_api_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get Tenant",
        "operationId": "get_tenant_api_t__tenant_slug__get",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tenants"
        ],
        "summary": "Update Tenant",
        "operationId": "update_tenant_api_t__tenant_slug__patch",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantSettingsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/archive": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Archive Tenant",
        "description": "회사를 보관한다. 데이터는 그대로 두고 읽기 전용으로 만들며, 일반 구성원에게는 숨긴다.",
        "operationId": "archive_tenant_api_t__tenant_slug__archive_post",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/restore": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Restore Tenant",
        "operationId": "restore_tenant_api_t__tenant_slug__restore_post",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/purge": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Purge Tenant",
        "description": "보관된 회사를 완전히 지운다. 문서·기록·감사 기록이 모두 사라지고 되돌릴 수 없다.\n\n지웠다는 사실(누가, 언제, 무엇을)은 deleted_tenant 에 남는다.",
        "operationId": "purge_tenant_api_t__tenant_slug__purge_post",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantPurgeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantPurgeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/org-units": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List Org Units",
        "operationId": "list_org_units_api_t__tenant_slug__org_units_get",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrgUnitOut"
                  },
                  "title": "Response List Org Units Api T  Tenant Slug  Org Units Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Create Org Unit",
        "operationId": "create_org_unit_api_t__tenant_slug__org_units_post",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/org-units/{unit_id}": {
      "patch": {
        "tags": [
          "tenants"
        ],
        "summary": "Update Org Unit",
        "operationId": "update_org_unit_api_t__tenant_slug__org_units__unit_id__patch",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tenants"
        ],
        "summary": "Delete Org Unit",
        "operationId": "delete_org_unit_api_t__tenant_slug__org_units__unit_id__delete",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/systems": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List Systems",
        "operationId": "list_systems_api_t__tenant_slug__systems_get",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemOut"
                  },
                  "title": "Response List Systems Api T  Tenant Slug  Systems Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Create System",
        "operationId": "create_system_api_t__tenant_slug__systems_post",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{tenant_slug}/systems/{system_slug}": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get System",
        "operationId": "get_system_api_t__tenant_slug__systems__system_slug__get",
        "parameters": [
          {
            "name": "system_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Slug"
            }
          },
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TenantIn": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "TenantIn"
      },
      "TenantOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "four_eyes": {
            "type": "boolean",
            "title": "Four Eyes"
          },
          "artifact_ai": {
            "type": "boolean",
            "title": "Artifact Ai"
          },
          "tenant_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Role"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "four_eyes",
          "artifact_ai",
          "tenant_role",
          "archived_at",
          "actions"
        ],
        "title": "TenantOut"
      },
...
    }
  }
}